+++ R/input-validation.R
@@ -63,9 +63,24 @@ input_validation_anticlustering <- function(x, K, objective, method,
 
   validate_input(
     method, "method", len = 1,
-    input_set = c("ilp", "exchange", "heuristic", "centroid", "local-maximum"), 
+    input_set = c("ilp", "exchange", "heuristic", "centroid", "local-maximum", "brusco"), 
     not_na = TRUE
   )
+  
+  if (method == "brusco") {
+    if (argument_exists(categories)) {
+      stop("It is not possible to use the algorithm by Brusco et al. with categorical restrictions.")
+    }
+    if (preclustering == TRUE) {
+      stop("It is not possible to use the algorithm by Brusco et al. with preclustering restrictions.")
+    }
+    if (length(K) != 1) {
+      stop("The argument `K` must indicate the number of groups when using `method = 'brusco'`.")
+    }
+    if (!objective %in% c("dispersion", "diversity")) {
+      stop("The algorithm by Brusco et al. can only be used to optimize diversity or dispersion.")
+    }
+  }
 
   if (method == "ilp") {
     check_if_solver_is_available()